DummyBot: Actually process lines from the server
We now relay special (commands in place of sender segment such as
ping/error), commands and numerics (i.e. 001) to specialised helpers,
which then determine the best course of action for dispatch.

Note, nothing really *happens* yet but we do deal with PING, and
autojoin the default channel when we gain our 001 numeric.