Add names support, switch to async-safe queue for output
Turns out using built-in async functionality for output is
not sane in glib, especially with multiple callbacks all
attempting to write to a pending socket. Now we just
immediately queue off the output, and as soon as we're
able to write some of the output, we will. This can
result in multiple messages being dispatched at a
time, in a safe ordered manner.