IrcTextWidget: Improve our rendering/handling of codes
Previously we added a single character at a time to the buffer, which was
time consuming and pretty dumb. Now we write the sanitized text to the buffer
after we have collected the rendering information for it in a set of iters
that will greatly increase the performance.

This works by counting the start and end points for each supported style,
and applying to the whole region at once, as opposed to adding a single char
with huge amounts of tags.