Introduce (not yet used) IrcParser
We'll need to start heavily validating our input from the IRC servers,
after it quickly became apparent that Quirky has the resolve of a slightly
defective fish against ircfuzz. First step is introducing a context based
validation parser, as with this commit. The next step will be to then
guard any usage of uninitialised parameters and such by blocking access
to methods via a set of requirements for each command/numeric.
To do this we'll use a two hashtables for callbacks, and set the requirements
as per the RFC, and drop (and report) invalid input.